+++ src/base/weather.py
@@ -1,8 +1,28 @@
+from typing import Any
 import httpx
 
 from database.funcs import cache
 from helpers.datatypes import WeatherData
 from config import config, logger
+
+
+def _get_coords_from_region(name: str) -> tuple[float, float]:
+    if f"coords_{name}" in cache:
+        logger.debug("cache")
+        return cache[f"coords_{name}"]
+
+    url = "https://geocoding-api.open-meteo.com/v1/search"
+
+    params = {
+        "name": name,
+        "count": 1,
+    }
+    response = httpx.get(url, params=params)
+    data: dict[str, list[dict[str, Any]]] = response.json()
+
+    result = data["results"][0]["latitude"], data["results"][0]["longitude"]
+    cache[f"coords_{name}"] = result
+    return result
 
 
 def get_weather() -> WeatherData:
@@ -11,12 +31,13 @@ def get_weather() -> WeatherData:
         return cache["weather_data"]
     logger.debug("new")
 
-    url = "http://api.openweathermap.org/data/2.5/weather?"
+    coords = _get_coords_from_region(config.weather.region)
+    url = "https://api.open-meteo.com/v1/forecast"
     params = {
-        "lang": "RU",
-        "q": config.weather.region,
-        "appid": config.weather.api_key,
-        "units": "metric",
+        "latitude": coords[0],
+        "longitude": coords[1],
+        "current": ["temperature_2m", "weather_code"],
+        "hourly": ["temperature_2m", "weather_code"],
     }
     response = httpx.get(url, params=params)
     weather_data = WeatherData(response.json())