+++ src/helpers/datatypes.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from typing import Any, Optional
 
 import transliterate
@@ -35,78 +36,131 @@ weather_emojis = {
     "Tornado": "🌪️",
 }
 
+wmo_weather_codes = {
+    0: "Clear",
+    1: "Clouds",
+    2: "Clouds",
+    3: "Clouds",
+    45: "Fog",
+    48: "Fog",
+    51: "Drizzle",
+    53: "Drizzle",
+    55: "Drizzle",
+    56: "Drizzle",
+    57: "Drizzle",
+    61: "Rain",
+    63: "Rain",
+    65: "Rain",
+    66: "Rain",
+    67: "Rain",
+    71: "Snow",
+    73: "Snow",
+    75: "Snow",
+    77: "Snow",
+    80: "Rain",
+    81: "Rain",
+    82: "Rain",
+    85: "Snow",
+    86: "Snow",
+    95: "Thunderstorm",
+    96: "Thunderstorm",
+    99: "Thunderstorm",
+}
 
-class Coordinates:
-    def __init__(self, coord_data: dict):
-        self.lon: float = coord_data.get("lon", 0.0)
-        self.lat: float = coord_data.get("lat", 0.0)
+wmo_weather_codes_ru = {
+    0: "Ясно",
+    1: "Преимущественно ясно",
+    2: "Переменная облачность",
+    3: "Пасмурно",
+    45: "Туман",
+    48: "Изморозь",
+    51: "Мелкий дождь",
+    53: "Умеренный дождь",
+    55: "Сильный дождь",
+    56: "Морозный дождь",
+    57: "Сильный морозный дождь",
+    61: "Слабый дождь",
+    63: "Умеренный дождь",
+    65: "Сильный дождь",
+    66: "Ледяной дождь",
+    67: "Сильный ледяной дождь",
+    71: "Слабый снегопад",
+    73: "Умеренный снегопад",
+    75: "Сильный снегопад",
+    77: "Снежные зерна",
+    80: "Ливень",
+    81: "Умеренный ливень",
+    82: "Сильный ливень",
+    85: "Слабый снег",
+    86: "Сильный снег",
+    95: "Гроза",
+    96: "Гроза с градом",
+    99: "Сильная гроза с градом",
+}
 
 
-class Weather:
-    def __init__(self, weather_data: dict):
-        self.id: int = weather_data.get("id", 0)
-        self.main: str = weather_data.get("main", "")
-        self.description: str = weather_data.get("description", "")
-        self.icon: str = weather_data.get("icon", "")
+class BaseWeather:
+    def __init__(self, weather_code: int):
+        self.weather_code: int = weather_code
 
     @property
-    def ru_name(self):
-        for _id, _weather_type in weather_types.items():
-            if str(self.id).startswith(_id):
-                return _weather_type.value
-        return ""
+    def type(self) -> str:
+        return wmo_weather_codes.get(self.weather_code, "Unknown")
 
     @property
-    def emoji(self):
-        return weather_emojis.get(str(self.main), "❓")
+    def emoji(self) -> str:
+        return weather_emojis.get(self.type, "❓")
+
+    @property
+    def ru_type(self) -> str:
+        return wmo_weather_codes_ru.get(self.weather_code, "Неизвестно")
 
 
-class MainInfo:
-    def __init__(self, main_data: dict):
-        self.temp: float = main_data.get("temp", 0.0)
-        self.feels_like: float = main_data.get("feels_like", 0.0)
-        self.temp_min: float = main_data.get("temp_min", 0.0)
-        self.temp_max: float = main_data.get("temp_max", 0.0)
-        self.pressure: int = main_data.get("pressure", 0)
-        self.humidity: int = main_data.get("humidity", 0)
+class HourlyWeather(BaseWeather):
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.time: list[datetime] = list(
+            map(lambda t: datetime.fromisoformat(t), data.get("time", []))
+        )
+        self.temperature_2m: list[float | int] = data.get("temperature_2m", [])
 
 
-class WindInfo:
-    def __init__(self, wind_data: dict):
-        self.speed: float = wind_data.get("speed", 0.0)
-        self.deg: float = wind_data.get("deg", 0.0)
+class HourlyUnits:
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.temperature_2m: str = data.get("temperature_2m", "°C")
+        self.time: str = data.get("time", "iso8601")
+        self.weather_code: str = data.get("weather_code", "wmo code")
 
 
-class CloudsInfo:
-    def __init__(self, clouds_data: dict):
-        self.all: int = clouds_data.get("all", 0)
+class CurrentWeather(BaseWeather):
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.time: datetime = datetime.fromisoformat(data.get("time", ""))
+        self.interval: int = data.get("interval", 0)
+        self.temperature_2m: float = data.get("temperature_2m", 0.0)
+        self.weather_code: int = data.get("weather_code", 0)
 
 
-class SysInfo:
-    def __init__(self, sys_data: dict):
-        self.type: int = sys_data.get("type", 0)
-        self.id: int = sys_data.get("id", 0)
-        self.country: str = sys_data.get("country", "")
-        self.sunrise: int = sys_data.get("sunrise", 0)
-        self.sunset: int = sys_data.get("sunset", 0)
+class CurrentUnits:
+    def __init__(self, data: dict[str, Any]):
+        self.time: str = data.get("time", "iso8601")
+        self.interval: int = data.get("interval", "seconds")
+        self.temperature_2m: float = data.get("temperature_2m", "°C")
+        self.weather_code: int = data.get("weather_code", "wmo code")
 
 
 class WeatherData:
     def __init__(self, data: dict[str, Any]):
-        self.data: dict = data
-        self.coord: Coordinates = Coordinates(data.get("coord", {}))
-        self.weather: Weather = Weather(data.get("weather", [{}])[0])
-        self.base: str = data.get("base", "")
-        self.main: MainInfo = MainInfo(data.get("main", {}))
-        self.visibility: int = data.get("visibility", 0)
-        self.wind: WindInfo = WindInfo(data.get("wind", {}))
-        self.clouds: CloudsInfo = CloudsInfo(data.get("clouds", {}))
-        self.dt: int = data.get("dt", 0)
-        self.sys: SysInfo = SysInfo(data.get("sys", {}))
-        self.timezone: int = data.get("timezone", 0)
-        self.id: int = data.get("id", 0)
-        self.name: str = data.get("name", "")
-        self.cod: int = data.get("cod", 0)
+        self.data: dict[str, Any] = data
+        self.latitude: float = data.get("latitude", 0.0)
+        self.longitude: float = data.get("latitude", 0.0)
+        self.elevation: float = data.get("elevation", 0.0)
+        self.generationtime_ms: float = data.get("generationtime_ms", 0.0)
+        self.utc_offset_seconds: int = data.get("utc_offset_seconds", 0)
+        self.timezone: str = data.get("timezone", "GMT")
+        self.timezone_abbreviation: str = data.get("timezone_abbreviation", "GMT")
+        self.hourly_units: HourlyUnits = HourlyUnits(data.get("hourly_units", {}))
+        self.hourly: HourlyWeather = HourlyWeather(data.get("hourly", {}))
+        self.current_units: CurrentUnits = CurrentUnits(data.get("current_units", {}))
+        self.current: CurrentWeather = CurrentWeather(data.get("current", {}))
 
 
 # ----------------------------------- item ----------------------------------- #